+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - FriendlyName parameter of Add-LabMachineDefinition now actually supported. Fixes #23
   - No interaction is done in AL using the friendly name. This is only for the purpose of
     deploying the same lab on the same host with different resource names
+- Including name of used function in telemetry, for all functions using Write-LogFunctionEntry
   
 ### Bug Fixes
 - Get-LabInternetFile did not work on Azure when the Uri did not contain a file name like 'https://go.microsoft.com/fwlink/?Linkid=85215'.